Dry-run option to test the configuration / scripts.

+++ src/aiori-POSIX.c
@@ -278,6 +278,9 @@ void *POSIX_Create(char *testFileName, IOR_param_t * param)
         if (param->useO_DIRECT == TRUE)
                 set_o_direct_flag(&fd_oflag);
 
+        if(param->dryRun)
+          return 0;
+
 #ifdef HAVE_LUSTRE_LUSTRE_USER_H
         if (param->lustre_set_striping) {
                 /* In the single-shared-file case, task 0 has to creat the
@@ -383,6 +386,10 @@ void *POSIX_Open(char *testFileName, IOR_param_t * param)
                 set_o_direct_flag(&fd_oflag);
 
         fd_oflag |= O_RDWR;
+
+        if(param->dryRun)
+          return 0;
+
         *fd = open64(testFileName, fd_oflag);
         if (*fd < 0)
                 ERR("open64 failed");
@@ -418,6 +425,9 @@ static IOR_offset_t POSIX_Xfer(int access, void *file, IOR_size_t * buffer,
         char *ptr = (char *)buffer;
         long long rc;
         int fd;
+
+        if(param->dryRun)
+          return length;
 
         fd = *(int *)file;
 
@@ -500,6 +510,8 @@ static void POSIX_Fsync(void *fd, IOR_param_t * param)
  */
 void POSIX_Close(void *fd, IOR_param_t * param)
 {
+        if(param->dryRun)
+          return;
         if (close(*(int *)fd) != 0)
                 ERR("close() failed");
         free(fd);
@@ -510,11 +522,14 @@ void POSIX_Close(void *fd, IOR_param_t * param)
  */
 void POSIX_Delete(char *testFileName, IOR_param_t * param)
 {
-        char errmsg[256];
-        sprintf(errmsg, "[RANK %03d]: unlink() of file \"%s\" failed\n",
-                rank, testFileName);
-        if (unlink(testFileName) != 0)
+        if(param->dryRun)
+          return;
+        if (unlink(testFileName) != 0){
+                char errmsg[256];
+                sprintf(errmsg, "[RANK %03d]: unlink() of file \"%s\" failed\n",
+                        rank, testFileName);
                 EWARN(errmsg);
+        }
 }
 
 /*
@@ -523,6 +538,8 @@ void POSIX_Delete(char *testFileName, IOR_param_t * param)
 IOR_offset_t POSIX_GetFileSize(IOR_param_t * test, MPI_Comm testComm,
                                       char *testFileName)
 {
+        if(test->dryRun)
+          return 0;
         struct stat stat_buf;
         IOR_offset_t aggFileSizeFromStat, tmpMin, tmpMax, tmpSum;
